Move shared components out
